몬스터-몬스터 충돌 추가, 넉백 수치 조정 필요

+++ ContentsProject/CharactersStatus.h
@@ -3,7 +3,7 @@
 
 struct CharacterStatus {
     std::string Name = "Imelda";
-    float Hp = 120.0f;
+    int Hp = 120;
     int Armor = 1;
     int Might = 0;
     int Speed = 160;
@@ -11,9 +11,9 @@ struct CharacterStatus {
 
     CharacterStatus() {};
 
-    CharacterStatus(std::string name, float hp, int armor, int might, int speed, int level = 1)
+    CharacterStatus(std::string name, int hp, int armor, int might, int speed, int level = 1)
         : Name(name), Hp(hp), Armor(armor), Might(might), Speed(speed), Level(level) {}
 };
                                 // Name, Health, Armor, Might, Speed
-inline const CharacterStatus AntonioStatus("Antonio", 120.0f, 1, 0, 160);
-inline const CharacterStatus ImeldaStatus("Imelda", 120.0f, 1, 0, 160);
+inline const CharacterStatus AntonioStatus("Antonio", 120, 1, 0, 160);
+inline const CharacterStatus ImeldaStatus("Imelda", 120, 1, 0, 160);
